Extract owner admin invites to separate route module
