+++ src/routes/owner/admins/handler.ts
@@ -1,7 +1,7 @@
 import { HttpStatus } from '@/net/http'
-import { getAdmin, getAdmins, inviteAdmin, resendAdminInvitation } from '@/use-cases/owner'
+import { getAdmin, getAdmins } from '@/use-cases/owner'
 
-import type { CreateInvitationCtx, GetAdminCtx, GetAdminsCtx, ResendInvitationCtx } from './schema'
+import type { GetAdminCtx, GetAdminsCtx } from './schema'
 
 export const getAdminsHandler = async (c: GetAdminsCtx) => {
   const { search, statuses, page, limit } = c.req.valid('query')
@@ -17,24 +17,6 @@ export const getAdminHandler = async (c: GetAdminCtx) => {
   const { adminId } = c.req.valid('param')
 
   const result = await getAdmin(adminId)
-
-  return c.json(result, HttpStatus.OK)
-}
-
-export const createInvitationHandler = async (c: CreateInvitationCtx) => {
-  const body = c.req.valid('json')
-  const { id: inviterId } = c.get('user')
-
-  const result = await inviteAdmin(body, inviterId)
-
-  return c.json(result, HttpStatus.CREATED)
-}
-
-export const resendInvitationHandler = async (c: ResendInvitationCtx) => {
-  const { adminId } = c.req.valid('param')
-  const { id: inviterId } = c.get('user')
-
-  const result = await resendAdminInvitation(adminId, inviterId)
 
   return c.json(result, HttpStatus.OK)
 }
