+++ src/routes/owner/admins/index.ts
@@ -4,18 +4,8 @@ import type { AppContext } from '@/context'
 
 import { requestValidator } from '@/middleware'
 
-import {
-  createInvitationHandler,
-  getAdminHandler,
-  getAdminsHandler,
-  resendInvitationHandler,
-} from './handler'
-import {
-  createInvitationBodySchema,
-  getAdminParamsSchema,
-  getAdminsQuerySchema,
-  resendInvitationParamsSchema,
-} from './schema'
+import { getAdminHandler, getAdminsHandler } from './handler'
+import { getAdminParamsSchema, getAdminsQuerySchema } from './schema'
 
 const ownerAdminsRoute = new Hono<AppContext>()
 
@@ -24,20 +14,11 @@ ownerAdminsRoute.get(
   requestValidator('query', getAdminsQuerySchema),
   getAdminsHandler,
 )
+
 ownerAdminsRoute.get(
   '/admins/:adminId',
   requestValidator('param', getAdminParamsSchema),
   getAdminHandler,
-)
-ownerAdminsRoute.post(
-  '/admins/invitations',
-  requestValidator('json', createInvitationBodySchema),
-  createInvitationHandler,
-)
-ownerAdminsRoute.post(
-  '/admins/invitations/:adminId/resend',
-  requestValidator('param', resendInvitationParamsSchema),
-  resendInvitationHandler,
 )
 
 export default ownerAdminsRoute
