+++ src/routes/owner/index.ts
@@ -1,1 +1,2 @@
 export { default as ownerAdminsRoute } from './admins'
+export { default as ownerAdminsInvitesRoute } from './admins/invites'
